+++ packages/transaction-controller/src/utils/batch.test.ts
@@ -1425,7 +1425,12 @@ describe('Batch Utils', () => {
             ...request,
             publishBatchHook: undefined,
             isSimulationEnabled: () => isSimulationSupportedMock(),
-            request: { ...request.request, useHook: true },
+            request: {
+              ...request.request,
+              disable7702: true,
+              disableHook: true,
+              disableSequential: false,
+            },
           }),
         ).rejects.toThrow(`Can't process batch`);
       });
@@ -1441,6 +1446,8 @@ describe('Batch Utils', () => {
             ...request.request,
             requireApproval: false,
             disable7702: true,
+            disableHook: true,
+            disableSequential: false,
           },
         }).catch(() => {
           // Intentionally empty
@@ -1464,7 +1471,12 @@ describe('Batch Utils', () => {
           addTransactionBatch({
             ...request,
             publishBatchHook: undefined,
-            request: { ...request.request, disable7702: true },
+            request: {
+              ...request.request,
+              disable7702: true,
+              disableHook: true,
+              disableSequential: false,
+            },
           }),
         ).rejects.toThrow('Test error');
 
@@ -1486,6 +1498,8 @@ describe('Batch Utils', () => {
             ...request.request,
             origin: ORIGIN_MOCK,
             disable7702: true,
+            disableHook: true,
+            disableSequential: false,
           },
         }).catch(() => {
           // Intentionally empty
@@ -1511,8 +1525,74 @@ describe('Batch Utils', () => {
 
         const result = await resultPromise;
         expect(result?.batchId).toMatch(/^0x[0-9a-f]{32}$/u);
+      });
+
+      it('updates gas properties', async () => {
+        const { approve } = mockRequestApproval(MESSENGER_MOCK, {
+          state: 'approved',
+        });
+        mockSequentialPublishBatchHookResults();
+        setupSequentialPublishBatchHookMock(() => sequentialPublishBatchHook);
+
+        const resultPromise = addTransactionBatch({
+          ...request,
+          publishBatchHook: undefined,
+          messenger: MESSENGER_MOCK,
+          request: {
+            ...request.request,
+            origin: ORIGIN_MOCK,
+            disable7702: true,
+            disableHook: true,
+          },
+        }).catch(() => {
+          // Intentionally empty
+        });
+
+        await flushPromises();
+        approve();
+        await executePublishHooks();
+
+        await resultPromise;
+
         expect(simulateGasBatchMock).toHaveBeenCalledTimes(1);
+        expect(simulateGasBatchMock).toHaveBeenCalledWith({
+          chainId: CHAIN_ID_MOCK,
+          from: FROM_MOCK,
+          transactions: [
+            {
+              params: TRANSACTION_BATCH_PARAMS_MOCK,
+            },
+            {
+              params: TRANSACTION_BATCH_PARAMS_MOCK,
+            },
+          ],
+        });
         expect(getGasFeesMock).toHaveBeenCalledTimes(1);
+        expect(getGasFeesMock).toHaveBeenCalledWith(
+          expect.objectContaining({
+            gasFeeControllerData: expect.any(Object),
+            messenger: MESSENGER_MOCK,
+            transactionMeta: {
+              chainId: CHAIN_ID_MOCK,
+              gas: GAS_TOTAL_MOCK,
+              from: FROM_MOCK,
+              networkClientId: NETWORK_CLIENT_ID_MOCK,
+              txParams: { from: FROM_MOCK, gas: GAS_TOTAL_MOCK },
+              origin: ORIGIN_MOCK,
+              id: expect.any(String),
+              status: 'unapproved',
+              time: expect.any(Number),
+              transactions: [
+                {
+                  params: TRANSACTION_BATCH_PARAMS_MOCK,
+                },
+                {
+                  params: TRANSACTION_BATCH_PARAMS_MOCK,
+                },
+              ],
+            },
+          }),
+        );
       });
 
       it('saves a transaction batch and then cleans the specific batch by ID', async () => {
@@ -1530,6 +1610,8 @@ describe('Batch Utils', () => {
             ...request.request,
             origin: ORIGIN_MOCK,
             disable7702: true,
+            disableHook: true,
+            disableSequential: false,
           },
         }).catch(() => {
           // Intentionally empty