+++ packages/transaction-controller/src/utils/batch.ts
@@ -73,28 +73,28 @@ type AddTransactionBatchRequest = {
   addTransaction: TransactionController['addTransaction'];
   getChainId: (networkClientId: string) => Hex;
   getEthQuery: (networkClientId: string) => EthQuery;
+  getGasFeeEstimates: (
+    options: FetchGasFeeEstimateOptions,
+  ) => Promise<GasFeeState>;
   getInternalAccounts: () => Hex[];
+  getPendingTransactionTracker: (
+    networkClientId: string,
+  ) => PendingTransactionTracker;
   getTransaction: (id: string) => TransactionMeta;
   isSimulationEnabled: () => boolean;
   messenger: TransactionControllerMessenger;
   publishBatchHook?: PublishBatchHook;
-  publicKeyEIP7702?: Hex;
-  request: TransactionBatchRequest;
-  updateTransaction: (
-    options: { transactionId: string },
-    callback: (transactionMeta: TransactionMeta) => void,
-  ) => void;
   publishTransaction: (
     _ethQuery: EthQuery,
     transactionMeta: TransactionMeta,
   ) => Promise<Hex>;
-  getPendingTransactionTracker: (
-    networkClientId: string,
-  ) => PendingTransactionTracker;
+  publicKeyEIP7702?: Hex;
+  request: TransactionBatchRequest;
   update: UpdateStateCallback;
-  getGasFeeEstimates: (
-    options: FetchGasFeeEstimateOptions,
-  ) => Promise<GasFeeState>;
+  updateTransaction: (
+    options: { transactionId: string },
+    callback: (transactionMeta: TransactionMeta) => void,
+  ) => void;
 };
 
 type IsAtomicBatchSupportedRequestInternal = {
@@ -433,9 +433,12 @@ async function addTransactionBatchWithHook(
     disableSequential = true;
   }
 
-  const publishBatchHook =
-    (!disableHook && requestPublishBatchHook) ??
-    (!disableSequential && sequentialPublishBatchHook.getHook());
+  let publishBatchHook = null;
+  if (!disableHook) {
+    publishBatchHook = requestPublishBatchHook;
+  } else if (!disableSequential) {
+    publishBatchHook = sequentialPublishBatchHook.getHook();
+  }
 
   if (!publishBatchHook) {
     log(`No supported batch methods found`, {
@@ -691,11 +694,10 @@ function wipeTransactionBatchById(
  * @returns A new TransactionBatchMeta object.
  */
 function newBatchMetadata(
-  transactionBatchMeta: Omit<TransactionBatchMeta, 'time' | 'status'>,
+  transactionBatchMeta: Omit<TransactionBatchMeta, 'status'>,
 ): TransactionBatchMeta {
   return {
     ...transactionBatchMeta,
-    time: Date.now(),
     status: TransactionStatus.unapproved,
   };
 }
@@ -732,6 +734,8 @@ async function prepareApprovalData({
     transactions: nestedTransactions,
   } = userRequest;
 
+  const ethQuery = getEthQuery(networkClientId);
+
   console.log('>>>>>>>> isSimulationEnabled', isSimulationEnabled());
   if (!isSimulationEnabled()) {
     console.log('throwing error');
@@ -764,7 +768,7 @@ async function prepareApprovalData({
   });
 
   const gasFeeResponse = await defaultGasFeeFlow.getGasFees({
-    ethQuery: getEthQuery(networkClientId),
+    ethQuery,
     gasFeeControllerData,
     messenger,
     transactionMeta: {
@@ -773,6 +777,7 @@ async function prepareApprovalData({
         from,
         gas: gasLimit,
       },
+      time: Date.now(),
     },
   });